+++ elks/include/linuxmt/sched.h
@@ -170,8 +170,8 @@ extern void kill_process(pid_t,sig_t,int);
 extern void add_to_runqueue(struct task_struct *);
 
 extern void arch_build_stack(struct task_struct *);
-extern unsigned int get_ustack(struct task_struct *,int);
-extern void put_ustack(register struct task_struct *,int,int);
+extern __u16 get_ustack(struct task_struct *,__u16);
+extern void put_ustack(register struct task_struct *,__u16,__u16);
 
 extern void tswitch(void);
